+++ internal/controller/utils.go
@@ -2,7 +2,7 @@ package controller
 
 import (
 	kyvernov1 "github.com/kyverno/kyverno/api/kyverno/v1"
-	kyvernov2alpha1 "github.com/kyverno/kyverno/api/kyverno/v2alpha1"
+	kyvernov2 "github.com/kyverno/kyverno/api/kyverno/v2"
 
 	giantswarmPolicy "github.com/giantswarm/kyverno-policy-operator/api/v1alpha1"
 )
@@ -24,10 +24,10 @@ func translateTargetsToResourceFilters(polex giantswarmPolicy.PolicyException) k
 }
 
 // translatePoliciesToExceptions takes a Kyverno ClusterPolicy array and transforms it into a Kyverno Exception array
-func translatePoliciesToExceptions(policies map[string]kyvernov1.ClusterPolicy) []kyvernov2alpha1.Exception {
-	var exceptionArray []kyvernov2alpha1.Exception
+func translatePoliciesToExceptions(policies map[string]kyvernov1.ClusterPolicy) []kyvernov2.Exception {
+	var exceptionArray []kyvernov2.Exception
 	for policyName, kyvernoPolicy := range policies {
-		kyvernoException := kyvernov2alpha1.Exception{
+		kyvernoException := kyvernov2.Exception{
 			PolicyName: policyName,
 			RuleNames:  generatePolicyRules(kyvernoPolicy),
 		}
@@ -51,7 +51,7 @@ func generatePolicyRules(kyvernoPolicy kyvernov1.ClusterPolicy) []string {
 }
 
 // unorderedEqual takes two Kyverno Exception arrays and checks if they are equal even if they are not ordered the same
-func unorderedEqual(got, want []kyvernov2alpha1.Exception) bool {
+func unorderedEqual(got, want []kyvernov2.Exception) bool {
 	// Check Length size first
 	if len(got) != len(want) {
 		return false