+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Update `Kyverno` API from v2beta1 to v2.
+
 ## [0.0.9] - 2025-05-21
 
 ### Changed